Optimization

+++ uninstall.sh
@@ -1,4 +1,12 @@
-rm -rf ~/.vim/autoload
-rm -rf ~/.vim/bundle
-rm -rf ~/.vim/colors
-rm ~/.vimrc
+if [ -f ~/.vim/autoload]; then
+    rm -rf ~/.vim/autoload
+fi
+if [ -f ~/.vim/bundle]; then
+    rm -rf ~/.vim/bundle
+fi
+if [ -f ~/.vim/colors]; then
+    rm -rf ~/.vim/colors
+fi
+if [ -f ~/.vimrc]; then
+    rm ~/.vimrc
+fi
